# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ISGIDocs API",
        "description": "ISGIDocs est une API conçue pour gérer les ressources pédagogiques et les formations de l'Institut Supérieur de Génie Informatique. Cette API permet aux utilisateurs de gérer les modules, les documents pédagogiques, et l'authentification des utilisateurs pour une expérience d'auto-formation fluide et organisée.",
        "contact": {
            "email": "salahaznidi09@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "Document"
                ],
                "summary": "Liste tous les documents",
                "operationId": "479e53af7d4820aecb8f9534fdbe21aa",
                "responses": {
                    "200": {
                        "description": "Liste des documents récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Document"
                ],
                "summary": "Ajoute un nouveau document",
                "operationId": "9b247d60e8834d470666fba8e773edb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "DOC101"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Document de Mathématiques 1"
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Introduction aux mathématiques"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Document sur les bases des mathématiques"
                                    },
                                    "chemin": {
                                        "type": "string",
                                        "example": "documents/math101.pdf"
                                    },
                                    "module_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données non valides"
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "get": {
                "tags": [
                    "Document"
                ],
                "summary": "Récupère un document par ID",
                "operationId": "0fa5befcdb9042b346d01baea49dee95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du document récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Document"
                ],
                "summary": "Modifie un document existant",
                "operationId": "b851458865e322163fe9b88595919255",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du document à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "DOC101"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Document de Mathématiques avancées"
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Introduction aux mathématiques avancées"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Document mis à jour pour le cours de mathématiques avancées"
                                    },
                                    "chemin": {
                                        "type": "string",
                                        "example": "documents/math101-updated.pdf"
                                    },
                                    "module_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Document"
                ],
                "summary": "Supprime un document",
                "operationId": "889c107c27dc6b118dee932d1760f574",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du document à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document supprimé avec succès"
                    },
                    "404": {
                        "description": "Document non trouvé"
                    }
                }
            }
        },
        "/api/documents/{id}/download": {
            "get": {
                "tags": [
                    "Document"
                ],
                "summary": "Télécharge un document",
                "operationId": "9c3894d9d7b13268e679345d76550898",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du document à télécharger",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Téléchargement du document réussi"
                    },
                    "404": {
                        "description": "Document non trouvé"
                    }
                }
            }
        },
        "/api/documents/search": {
            "get": {
                "tags": [
                    "Document"
                ],
                "summary": "Rechercher des documents",
                "operationId": "44b72b64ecb823090925fe2ab6326a2b",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Terme de recherche pour le nom ou la description du document",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documents trouvés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Aucun terme de recherche fourni"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Document": {
                "title": "Document",
                "description": "Schéma pour un document",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Introduction aux bases de données"
                    },
                    "type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "url": {
                        "type": "string",
                        "example": "http://example.com/document.pdf"
                    },
                    "module_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer un token d'accès après avoir été authentifié.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}